+++ glennifer/automodule/src/main/java/com/cwrubotix/glennifer/automodule/Obstacle.java
@@ -2,15 +2,15 @@ package com.cwrubotix.glennifer.automodule;
 
 public class Obstacle extends Coordinate {
 
-    private float radius;
-    private final float KINECT_ERROR_BOUND = 0.05F;
+    private double radius;
+    private final double KINECT_ERROR_BOUND = 0.05F;
 
-    public Obstacle(float x_pos, float y_pos, float radius) {
+    public Obstacle(double x_pos, double y_pos, double radius) {
         super(x_pos, y_pos);
         this.radius = radius;
     }
 
-    public float getRadius() {
+    public double getRadius() {
         return radius;
     }
 
@@ -18,7 +18,7 @@ public class Obstacle extends Coordinate {
         this(obs.getX(), obs.getY(), 0.15F);
     }
 
-    public void setRadius(float radius) {
+    public void setRadius(double radius) {
         this.radius = radius;
     }
     
@@ -26,9 +26,9 @@ public class Obstacle extends Coordinate {
     public boolean equals(Object obj){
 	if(obj instanceof Obstacle){
 	    Obstacle other = (Obstacle)obj;
-	    float x_diff = getX() - other.getX();
-	    float y_diff = getY() - other.getY();
-	    float r_diff = getRadius() - other.getRadius();
+	    double x_diff = getX() - other.getX();
+	    double y_diff = getY() - other.getY();
+	    double r_diff = getRadius() - other.getRadius();
 	    return x_diff <= KINECT_ERROR_BOUND && y_diff <= KINECT_ERROR_BOUND && r_diff <= KINECT_ERROR_BOUND;
 	}
 	return false;
